separating parallel module
