+++ main.py
@@ -6,20 +6,9 @@ from diffusers.pipelines.stable_diffusion.safety_checker import (
     StableDiffusionSafetyChecker,
 )
 import os
-from utils import StableDiffusionMultiProcessing, get_gpu_setting, dummy_checker
+from utils import get_gpu_setting, dummy_checker
+from parallel import StableDiffusionMultiProcessing
 from schedulers import schedulers
-
-
-def image_grid(imgs, rows, cols):
-    assert len(imgs) == rows * cols
-
-    w, h = imgs[0].size
-    grid = Image.new("RGB", size=(cols * w, rows * h))
-    grid_w, grid_h = grid.size
-
-    for i, img in enumerate(imgs):
-        grid.paste(img, box=(i % cols * w, i // cols * h))
-    return grid
 
 
 TOKEN = os.environ.get("TOKEN", None)
@@ -44,13 +33,13 @@ kwargs = dict(
 
 if multi:
     # "data parallel", replicate the model on multiple gpus, each is handled by a different process
-    pipe: StableDiffusionMultiProcessing = StableDiffusionMultiProcessing.from_pretrained(
-        devices, **kwargs
-    )
+    pipe = StableDiffusionMultiProcessing.from_pretrained(devices, **kwargs)
+    
 else:
     pipe: StableDiffusionPipeline = StableDiffusionPipeline.from_pretrained(**kwargs)
     # remove safety checker so it doesn't use up GPU memory
     safety: StableDiffusionSafetyChecker = pipe.safety_checker
+    # TODO remove CLIP feature extractor, pre-processing step for safety checker
     pipe.safety_checker = dummy_checker
     if len(devices):
         pipe.to(f"cuda:{devices[0]}")
@@ -109,6 +98,8 @@ def inference(
 
 
 if __name__ == "__main__":
+    from utils import image_grid
+
     images = inference(input("Input prompt:"))
     grid = image_grid(images, rows=1, cols=1)
     grid.show()
